first debugging commit

+++ Node.cpp
@@ -13,14 +13,14 @@ private:
     int movieLengthMinutes;
     int movieReleaseYear;
     string movieGenre;
-    float movieRating;
+    int movieRating;
 
 public:
     Node *next;
 
     Node() : next(nullptr) {}
 
-    Node(string _movieTitle, int _movieLengthMinutes, int _movieReleaseYear, string _movieGenre, float _movieRating)
+    Node(string _movieTitle, int _movieLengthMinutes, int _movieReleaseYear, string _movieGenre, int _movieRating)
     {
         movieTitle = _movieTitle;
         movieLengthMinutes = _movieLengthMinutes;
@@ -33,11 +33,11 @@ public:
     void setLength(int _length) { movieLengthMinutes = _length; }
     void setReleaseYear(int _releaseYear) { movieReleaseYear = _releaseYear; }
     void setGenre(string _genre) { movieGenre = _genre; }
-    void setRating(float _rating) { movieRating = _rating; }
+    void setRating(int _rating) { movieRating = _rating; }
 
     string getTitle() const { return movieTitle; }
     int getLength() const { return movieLengthMinutes; }
     int getReleaseYear() const { return movieReleaseYear; }
     string getGenre() const { return movieGenre; }
-    float getRating() const { return movieRating; }
+    int getRating() const { return movieRating; }
 };
